[PromotionBundle] Split PromotionCouponGeneratorInstructionInterface class into separate getter and setter interface
