+++ src/Sylius/Component/Promotion/Exception/FailedGenerationException.php
@@ -13,12 +13,12 @@ declare(strict_types=1);
 
 namespace Sylius\Component\Promotion\Exception;
 
-use Sylius\Component\Promotion\Generator\PromotionCouponGeneratorInstructionInterface;
+use Sylius\Component\Promotion\Generator\PromotionCouponGeneratorInstructionReadInterface;
 
 final class FailedGenerationException extends \InvalidArgumentException
 {
     public function __construct(
-        PromotionCouponGeneratorInstructionInterface $instruction,
+        PromotionCouponGeneratorInstructionReadInterface $instruction,
         int $exceptionCode = 0,
         ?\Exception $previousException = null,
     ) {
