+++ src/Sylius/Component/Promotion/Generator/PromotionCouponGeneratorInstructionInterface.php
@@ -13,32 +13,20 @@ declare(strict_types=1);
 
 namespace Sylius\Component\Promotion\Generator;
 
-interface PromotionCouponGeneratorInstructionInterface
+interface PromotionCouponGeneratorInstructionInterface extends PromotionCouponGeneratorInstructionReadInterface
 {
-    public function getAmount(): ?int;
-
     public function setAmount(?int $amount): void;
-
-    public function getPrefix(): ?string;
 
     public function setPrefix(?string $prefix): void;
 
-    public function getCodeLength(): ?int;
-
     public function setCodeLength(?int $codeLength): void;
 
-    public function getSuffix(): ?string;
-
     public function setSuffix(?string $suffix): void;
-
-    public function getExpiresAt(): ?\DateTimeInterface;
 
     /**
      * @param \DateTimeInterface $expiresAt
      */
     public function setExpiresAt(?\DateTimeInterface $expiresAt): void;
-
-    public function getUsageLimit(): ?int;
 
     public function setUsageLimit(int $usageLimit): void;
 }
