+++ src/Sylius/Component/Promotion/spec/Generator/PercentageGenerationPolicySpec.php
@@ -15,7 +15,7 @@ namespace spec\Sylius\Component\Promotion\Generator;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Promotion\Generator\GenerationPolicyInterface;
-use Sylius\Component\Promotion\Generator\PromotionCouponGeneratorInstructionInterface;
+use Sylius\Component\Promotion\Generator\PromotionCouponGeneratorInstructionReadInterface;
 use Sylius\Component\Promotion\Repository\PromotionCouponRepositoryInterface;
 
 final class PercentageGenerationPolicySpec extends ObjectBehavior
@@ -31,7 +31,7 @@ final class PercentageGenerationPolicySpec extends ObjectBehavior
     }
 
     function it_examines_possibility_of_coupon_generation(
-        PromotionCouponGeneratorInstructionInterface $instruction,
+        PromotionCouponGeneratorInstructionReadInterface $instruction,
         PromotionCouponRepositoryInterface $couponRepository,
     ): void {
         $instruction->getAmount()->willReturn(17);
@@ -44,7 +44,7 @@ final class PercentageGenerationPolicySpec extends ObjectBehavior
     }
 
     function it_examines_possibility_of_coupon_generation_with_prefix_and_suffix(
-        PromotionCouponGeneratorInstructionInterface $instruction,
+        PromotionCouponGeneratorInstructionReadInterface $instruction,
         PromotionCouponRepositoryInterface $couponRepository,
     ): void {
         $instruction->getAmount()->willReturn(7);
@@ -60,7 +60,7 @@ final class PercentageGenerationPolicySpec extends ObjectBehavior
     }
 
     function it_returns_possible_generation_amount(
-        PromotionCouponGeneratorInstructionInterface $instruction,
+        PromotionCouponGeneratorInstructionReadInterface $instruction,
         PromotionCouponRepositoryInterface $couponRepository,
     ): void {
         $instruction->getAmount()->willReturn(17);
@@ -73,7 +73,7 @@ final class PercentageGenerationPolicySpec extends ObjectBehavior
     }
 
     function it_returns_php_int_max_value_as_possible_generation_amount_when_code_length_is_too_large(
-        PromotionCouponGeneratorInstructionInterface $instruction,
+        PromotionCouponGeneratorInstructionReadInterface $instruction,
         PromotionCouponRepositoryInterface $couponRepository,
     ): void {
         $instruction->getAmount()->willReturn(1000);
@@ -86,7 +86,7 @@ final class PercentageGenerationPolicySpec extends ObjectBehavior
     }
 
     function it_returns_possible_generation_amount_with_prefix_and_suffix(
-        PromotionCouponGeneratorInstructionInterface $instruction,
+        PromotionCouponGeneratorInstructionReadInterface $instruction,
         PromotionCouponRepositoryInterface $couponRepository,
     ): void {
         $instruction->getAmount()->willReturn(3);
@@ -102,7 +102,7 @@ final class PercentageGenerationPolicySpec extends ObjectBehavior
     }
 
     function it_throws_an_invalid_argument_exception_when_expected_amount_is_null(
-        PromotionCouponGeneratorInstructionInterface $instruction,
+        PromotionCouponGeneratorInstructionReadInterface $instruction,
     ): void {
         $instruction->getAmount()->willReturn(null);
         $instruction->getCodeLength()->willReturn(1);
@@ -112,7 +112,7 @@ final class PercentageGenerationPolicySpec extends ObjectBehavior
     }
 
     function it_throws_an_invalid_argument_exception_when_expecte_code_length_is_null(
-        PromotionCouponGeneratorInstructionInterface $instruction,
+        PromotionCouponGeneratorInstructionReadInterface $instruction,
     ): void {
         $instruction->getAmount()->willReturn(18);
         $instruction->getCodeLength()->willReturn(null);
